New Feature: zt_atexit
zt_atexit is an attempt to overcome the limits of the standard atexit
the API zt_atexit(void (*fn)(void * data), void * data) allows the
registrar to pass data to the callback (fn).  There is no limit (other
then memory) of the number of exit callbacks that can be registered.

Currently the functions are called in LIFO order (but this should not be
relied upon).