Fix formatter bug on linux
We were improperly using va_arg in sub-functions and expecting the
result to be normal.  per specification: "If ap is passed to a function
that uses va_arg(ap,type) then the value of ap is undefined after the
return of that function."

In theory we could have passed a pointer to va_list and it would work
(noted in C99) however this is also not portable.

Moved to a look table and void * instead.

Also fixed a size_t bug where we were assuming unsigned long size
pointers when printing them.