Convert zt_opt to use callbacks for all cases
zt_opt attempted to do too much fiddling with the types
being passed to it.  This version converts the "static"
types with function callbacks.  The builtin callback
types will read the named type from the arg list, run
sanity checks against the string and convert it to the
named type.