Updated the table API to support arbitrary data
Updated the key and value portions of the table API to
support key length and data length.  It is probably exposed in more
places then necessary but due to the callback nature of it I decided it
was best.

Also fixed a bug in the set logic (related to the size fixes in tables)