Added zt_uuid_base62_hashable_fmt print format
The base 62 algorithm being used leaves very little of the 64 bits being
consumed avaialble when generating the first 2 bytes of the output format;
This hashes of the values will cluster in an unreasonably small space (00 - lz).

This commit adds a new target format zt_uuid_base62_hashable_fmt that
mirrors the first 80 bits of the print format (e.g. 1234 -> 4321) so
that the limited space is moved towards the center of the print format.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch feature/base62_bits
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   src/zt_uuid.c
#	modified:   src/zt_uuid.h
#	modified:   tests/uuid_test.c
#
# Changes not staged for commit:
#   (use "git add <file>..." to update what will be committed)
#   (use "git checkout -- <file>..." to discard changes in working directory)
#
#	modified:   CMakeModules/asciidoc.cmake
#	modified:   README
#	modified:   doc/CMakeLists.txt
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	.build/
#	.syntastic_c_config
#	doc/man/CMakeLists.txt
#	doc/man/manpages.conf
#	uuid_test
#	uuid_test.c